+++ docs/history.rst
@@ -5,6 +5,7 @@ Changelog
 ------
 - Update on channel and connection callback to accept any `*args, **kwargs` to accomodate Pika updates. Since arguments are not used, except logging only.
 - Update unit tests that may be broken due to asyncio / tornado updates.
+- Shutdown message connection it we detect it's not running to allow the MCP to attempt reconnect.
 
 3.22.3
 ------